app/requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
Django>=3.0,<4.0
psycopg2-binary>=2.8
djangorestframework>=3.12.4
Pillow>=8.2.0
djangorestframework-simplejwt==4.6.0